Move content module loading and resolution to the content module loader
