+++ intellij-plugin-structure/structure-intellij/src/main/java/com/jetbrains/plugin/structure/intellij/plugin/loaders/ContentModuleLoader.kt
@@ -4,18 +4,30 @@
 
 package com.jetbrains.plugin.structure.intellij.plugin.loaders
 
+import com.jetbrains.plugin.structure.base.plugin.PluginCreationFail
+import com.jetbrains.plugin.structure.base.plugin.PluginCreationResult
 import com.jetbrains.plugin.structure.base.plugin.PluginCreationSuccess
+import com.jetbrains.plugin.structure.base.problems.PluginProblem
+import com.jetbrains.plugin.structure.base.problems.PluginProblem.Level.ERROR
 import com.jetbrains.plugin.structure.base.problems.UnableToReadDescriptor
 import com.jetbrains.plugin.structure.base.utils.isJar
 import com.jetbrains.plugin.structure.base.utils.toSystemIndependentName
+import com.jetbrains.plugin.structure.intellij.plugin.IdePlugin
+import com.jetbrains.plugin.structure.intellij.plugin.IdePluginImpl
+import com.jetbrains.plugin.structure.intellij.plugin.InlineDeclaredModuleV2Dependency
 import com.jetbrains.plugin.structure.intellij.plugin.Module.FileBasedModule
 import com.jetbrains.plugin.structure.intellij.plugin.Module.InlineModule
+import com.jetbrains.plugin.structure.intellij.plugin.ModuleDescriptor
+import com.jetbrains.plugin.structure.intellij.plugin.ModuleLoadingRule
 import com.jetbrains.plugin.structure.intellij.plugin.PluginCreator
 import com.jetbrains.plugin.structure.intellij.plugin.PluginCreator.Companion.createInvalidPlugin
 import com.jetbrains.plugin.structure.intellij.plugin.PluginCreator.Companion.createPlugin
+import com.jetbrains.plugin.structure.intellij.plugin.PluginDependency
 import com.jetbrains.plugin.structure.intellij.plugin.PluginLoader
 import com.jetbrains.plugin.structure.intellij.plugin.descriptors.DescriptorResource
 import com.jetbrains.plugin.structure.intellij.problems.AnyProblemToWarningPluginCreationResultResolver
+import com.jetbrains.plugin.structure.intellij.problems.ModuleDescriptorProblem
+import com.jetbrains.plugin.structure.intellij.problems.ModuleDescriptorResolutionProblem
 import com.jetbrains.plugin.structure.intellij.problems.PluginCreationResultResolver
 import com.jetbrains.plugin.structure.intellij.resources.ResourceResolver
 import com.jetbrains.plugin.structure.intellij.utils.JDOMUtil
@@ -67,6 +79,87 @@ class ContentModuleLoader internal constructor(private val pluginLoader: PluginL
     currentPlugin.addModuleDescriptor(module, module.loadingRule, moduleDescriptorResource, moduleCreator)
   }
 
+  internal fun PluginCreator.addModuleDescriptor(
+    moduleName: String,
+    loadingRule: ModuleLoadingRule,
+    configurationFile: String,
+    moduleCreator: PluginCreator
+  ) {
+    val pluginCreationResult = moduleCreator.pluginCreationResult
+    if (pluginCreationResult is PluginCreationSuccess<IdePlugin>) {
+      val module = pluginCreationResult.plugin
+
+      plugin.addDependencies(module, loadingRule)
+      plugin.modulesDescriptors.add(
+        ModuleDescriptor(
+          moduleName,
+          loadingRule,
+          module.dependencies,
+          module,
+          configurationFile
+        )
+      )
+      plugin.definedModules.add(moduleName)
+
+      mergeContent(module)
+    } else {
+      registerProblem(ModuleDescriptorResolutionProblem(moduleName, configurationFile, pluginCreationResult.errors))
+    }
+  }
+
+  internal fun PluginCreator.addModuleDescriptor(
+    inlineModuleReference: InlineModule,
+    loadingRule: ModuleLoadingRule,
+    moduleDescriptorResource: DescriptorResource,
+    moduleCreator: PluginCreator
+  ) {
+    val pluginCreationResult = moduleCreator.pluginCreationResult
+    if (pluginCreationResult is PluginCreationSuccess<IdePlugin>) {
+      val moduleName = inlineModuleReference.name
+      val inlineModule = pluginCreationResult.plugin
+
+      plugin.addInlineModuleDependencies(inlineModuleReference, inlineModule, loadingRule)
+      plugin.modulesDescriptors.add(
+        ModuleDescriptor.of(
+          moduleName,
+          loadingRule,
+          inlineModule,
+          moduleDescriptorResource
+        )
+      )
+      plugin.definedModules.add(moduleName)
+
+      mergeContent(inlineModule)
+    } else {
+      registerProblem(ModuleDescriptorProblem(inlineModuleReference, pluginCreationResult.errors))
+    }
+  }
+
+  private fun IdePluginImpl.addDependencies(module: IdePlugin, loadingRule: ModuleLoadingRule) {
+    module.forEachDependencyNotIn(this) {
+      val dependency = if (loadingRule.required) it else it.asOptional()
+      dependencies += dependency
+    }
+  }
+
+  private fun IdePluginImpl.addInlineModuleDependencies(
+    inlineModuleReference: InlineModule,
+    module: IdePlugin,
+    loadingRule: ModuleLoadingRule
+  ) {
+    module.forEachDependencyNotIn(this) {
+      dependencies += InlineDeclaredModuleV2Dependency.of(it.id, loadingRule, contentModuleOwner = this, inlineModuleReference)
+    }
+  }
+
+
+  private fun IdePlugin.forEachDependencyNotIn(plugin: IdePlugin, dependencyHandler: (PluginDependency) -> Unit) {
+    return dependencies
+      .filter { dependency -> plugin.dependencies.none { it.id == dependency.id } }
+      .forEach { dependencyHandler(it) }
+  }
+
+
   private fun getDescriptorResource(module: InlineModule, pluginFile: Path, descriptorPath: String): DescriptorResource {
     // TODO descriptor path is not relative to the pluginFile JAR. See MP-7224
     val parentUriStr = if (pluginFile.isJar()) {
@@ -94,8 +187,8 @@ class ContentModuleLoader internal constructor(private val pluginLoader: PluginL
           descriptorXml,
           resourceResolver,
           problemResolver
-        ).also {
-          logPluginCreationWarnings(moduleId, it)
+        ).also { creator ->
+          logPluginCreationWarnings(moduleId, creator)
         }
       } catch (e: IOException) {
         with(descriptorResource) {
@@ -117,4 +210,9 @@ class ContentModuleLoader internal constructor(private val pluginLoader: PluginL
     }
   }
 
+  private val PluginCreationResult<IdePlugin>.errors: List<PluginProblem>
+    get() = when (this) {
+      is PluginCreationSuccess -> emptyList()
+      is PluginCreationFail -> this.errorsAndWarnings.filter { it.level === ERROR }
+    }
 }
